新增识别系统为ubuntu还是centos

+++ setup.sh
@@ -1,9 +1,14 @@
+#!/bin/sh
+# 安装新开服务器的初始环境，当然最好是装完保存封装为一个基础系统予以复用。
+
+centosReleasePath="/etc/redhat-release"
 echo '安装Java'
 tar vxf jdk-8u171-linux-x64.tar.gz
 mv ./jdk1.8.0_171 /usr/lib/jdk1.8.0_171
 echo 'export JAVA_HOME=/usr/lib/jdk1.8.0_171'>>/etc/profile
 echo 'export CLASSPATH=.:$JAVA_HOME/lib:$JAVA_HOME/jre/lib:$CLASSPATH'>>/etc/profile
 echo 'export PATH=$JAVA_HOME/bin:$JAVA_HOME/jre/bin:$PATH'>>/etc/profile
+
 
 echo '安装Maven'
 tar vxf apache-maven-3.5.3-bin.tar.gz
@@ -15,18 +20,28 @@ echo '安装Gradle'
 wget https://downloads.gradle.org/distributions/gradle-4.9-bin.zip
 mkdir /opt/gradle
 # 安装unzip，ubuntu
-apt-get install unzip
-# centos:yum install zip unzip
+# 如果找不到centos的资源文件，说明默认是ubuntu系统
+if [ ! -f "$centosReleasePath" ]; 
+then
+	apt-get install unzip
+else 
+	yum install zip unzip
+fi
+
 unzip -d /opt/gradle gradle-4.9-bin.zip
 echo 'export PATH=$PATH:/opt/gradle/gradle-4.9/bin'>>/etc/profile
 
 echo '安装docker'  
-#CentOS安装命令 
-#sudo yum update
-#sudo yum install docker
+if [ ! -f "$centosReleasePath" ]; 
+then
 #Ubuntu安装命令
-#sudo apt-get update
-#sudo apt-get  install -y  docker.io
+	sudo apt-get update
+	sudo apt-get  install -y  docker.io
+else
+#CentOS安装命令 
+	sudo yum update
+	sudo yum install docker
+fi
 
 echo '安装docker-compose'
 mv ./docker-compose /usr/local/bin/docker-compose
